+++ src/test/java/org/cups4j/operations/AbstractIppOperationTest.java
@@ -1,17 +1,16 @@
 /*
- * Copyright (c) 2018 by Oliver Boehm
+ * Copyright (c) 2018-2026 by Oliver Boehm
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ * This program is free software; you can redistribute it and/or modify it under the terms of the
+ * GNU Lesser General Public License as published by the Free Software Foundation; either version 3
+ * of the License, or (at your option) any later version.
  *
- *   http://www.apache.org/licenses/LICENSE-2.0
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without
+ * even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * See the GNU Lesser General Public License for more details. You should have received a copy of
+ * the GNU Lesser General Public License along with this program; if not, see
+ * <http://www.gnu.org/licenses/>.
  *
  * (c)reated 23.03.2018 by oboehm (ob@oasd.de)
  */
@@ -35,6 +34,8 @@ import org.testcontainers.utility.DockerImageName;
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
 import java.net.MalformedURLException;
+import java.net.URI;
+import java.net.URISyntaxException;
 import java.net.URL;
 import java.nio.ByteBuffer;
 import java.util.HashMap;
@@ -64,7 +65,7 @@ public abstract class AbstractIppOperationTest {
     protected CupsClient client;
    
     protected ByteBuffer getIppHeader(IppOperation operation) throws UnsupportedEncodingException {
-        URL printerURL = this.getPrinterURL();
+        URI printerURL = this.getPrinterURL();
         Map<String, String> attributes = setUpAttributes();
         return operation.getIppHeader(printerURL, attributes);
     }
@@ -93,10 +94,10 @@ public abstract class AbstractIppOperationTest {
         }
     }
     
-    protected URL getPrinterURL() {
+    protected URI getPrinterURL() {
         try {
-            return new URL("http", cups.getHost(), cups.getFirstMappedPort(), "/printers/" + this.printerName);
-        } catch (MalformedURLException ex) {
+            return new URL("http", cups.getHost(), cups.getFirstMappedPort(), "/printers/" + this.printerName).toURI();
+        } catch (MalformedURLException | URISyntaxException ex) {
             throw new IllegalArgumentException("Invalid URL: ", ex);
         }
     }